Updated the entities
