+++ backend/VehicleManagementSystem/src/main/java/com/vms/entities/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.VehicleManagementSystem.entities;
+package com.vms.entities;
 
 import java.time.LocalDate;
