+++ backend/VehicleManagementSystem/src/main/java/com/vms/entities/Company.java
@@ -1,4 +1,4 @@
-package com.VehicleManagementSystem.entities;
+package com.vms.entities;
 
 
 import jakarta.persistence.*;
