+++ backend/VehicleManagementSystem/src/main/java/com/vms/entities/Payment.java
@@ -1,4 +1,4 @@
-package com.VehicleManagementSystem.entities;
+package com.vms.entities;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -9,8 +9,8 @@ import lombok.Setter;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-import com.VehicleManagementSystem.entities.enums.PaymentMethod;
-import com.VehicleManagementSystem.entities.enums.PaymentStatus;
+import com.vms.entities.enums.PaymentMethod;
+import com.vms.entities.enums.PaymentStatus;
 
 @Entity
 @Getter
