+++ backend/VehicleManagementSystem/src/main/java/com/vms/entities/Review.java
@@ -1,4 +1,4 @@
-package com.VehicleManagementSystem.entities;
+package com.vms.entities;
 
 import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.Column;
