+++ backend/VehicleManagementSystem/src/main/java/com/vms/entities/Variant.java
@@ -1,7 +1,7 @@
-package com.VehicleManagementSystem.entities;
+package com.vms.entities;
 
 
-import com.VehicleManagementSystem.entities.enums.FuelType;
+import com.vms.entities.enums.FuelType;
 
 import jakarta.persistence.*;
 import lombok.*;
