+++ backend/VehicleManagementSystem/src/main/java/com/vms/entities/Vehicle.java
@@ -1,0 +1,50 @@
+package com.vms.entities;
+
+import com.vms.entities.enums.*;
+import jakarta.persistence.*;
+import lombok.*;
+import java.math.BigDecimal;
+import java.time.LocalDate;
+import java.util.List;
+
+@Entity
+@Table(name = "vehicle")
+@NoArgsConstructor
+@AllArgsConstructor
+public class Vehicle extends BaseEntity {
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "variant_id", nullable = false)
+    private Variant variant;
+
+    @Column(name = "registration_number", length = 50, unique = true, nullable = false)
+    private String registrationNumber;
+
+    @Column(length = 50)
+    private String color;
+
+    @Lob
+    private String image;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "availability_status", length = 20)
+    private AvailabilityStatus availabilityStatus = AvailabilityStatus.AVAILABLE;
+
+    @Column(name = "price_per_day", precision = 10, scale = 2)
+    private BigDecimal pricePerDay;
+
+    @Column(name = "mileage", nullable = false)
+    private int mileage;
+
+    @Column(name = "updated_on")
+    private LocalDate updatedOn;
+
+    private String x1;
+    private String x2;
+
+    @OneToMany(mappedBy = "vehicle", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Review> reviews;
+
+    @OneToMany(mappedBy = "vehicle", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Booking> bookings;
+}
