+++ backend/VehicleManagementSystem/src/main/java/com/vms/entities/enums/BookingStatus.java
@@ -1,4 +1,4 @@
-package com.VehicleManagementSystem.entities.enums;
+package com.vms.entities.enums;
 
 public enum BookingStatus {
     PENDING, APPROVED, CONFIRMED, CANCELLED
