+++ backend/VehicleManagementSystem/src/main/java/com/vms/entities/enums/FuelType.java
@@ -1,4 +1,4 @@
-package com.VehicleManagementSystem.entities.enums;
+package com.vms.entities.enums;
 
 public enum FuelType {
 	    PETROL,
